isUserModerator extracted away to a new module moderate
